requirements.txt
aiomysql==0.0.20
aiosqlite==0.13.0
antiorm==1.2.1
asyncpg==0.20.1
certifi==2020.4.5.2
cffi==1.14.0
colorama==0.4.3
cryptography==2.9.2
db==0.1.1
iso8601==0.1.12
pycparser==2.20
PyMySQL==0.9.2
PyPika==0.37.8
python-dotenv==0.13.0
redis==3.5.3
sentry-sdk==0.15.1
six==1.15.0
tortoise-orm==0.16.13
typing-extensions==3.7.4.2
urllib3==1.25.9
